Move commands module out of frontend package
