+++ neurolang/frontend/deterministic_frontend.py
@@ -20,7 +20,7 @@ from .datalog.sugar import (
     TranslateSelectByFirstColumn,
     TranslateHeadConstantsToEqualities
 )
-from .commands import CommandsMixin
+from ..commands import CommandsMixin
 from .frontend_extensions import NumpyFunctionsMixin
 from .query_resolution_datalog import QueryBuilderDatalog
 from .query_resolution_expressions import Symbol
